+++ apps/webapp/shared/filled-image.tsx
@@ -1,13 +1,13 @@
 import Image from 'next/image';
 import { cn } from 'ui/lib/utils';
 
-interface ResponsiveImageProps {
+interface FilledImageProps {
   src: string;
   alt: string;
   className?: string;
 }
 
-export const ImageWrapper = ({ src, alt, className }: ResponsiveImageProps) => {
+export const FilledImage = ({ src, alt, className }: FilledImageProps) => {
   return (
     <div className={cn('relative', className)}>
       <Image
