+++ apps/webapp/shared/index.tsx
@@ -1,5 +1,5 @@
 export * from './edu-info-card';
-export * from './image-wrapper';
+export * from './filled-image';
 export * from './input-block';
 export * from './input-token';
-export * from './components/unoptimized-image';
+export * from './unoptimized-image';
